Rooms divided, visitors generation changed, day time mechanics added, potions cabinet mechanicks changed and theirs upgrades removed.
I divided the rooms according to the type of illness, completely changed the system for the arrival of patients - I did it by day, and not by purchased machines, the change of day and night (depending on the proximity to sunset - changing the delay in the arrival of patients). I removed the mechanics of upgrading the potion cabinet, reworked the cabinet so that it accepts each potion, 2 pieces per shelf, according to the type of elixir.